Moved pages into their own directory
