+++ src/pages/Login.vue
@@ -13,7 +13,7 @@
 
 <script>
 import { mapActions } from 'vuex';
-import Button from './Button.vue';
+import Button from '../components/Button.vue';
 
 export default {
     components: {
